Moved Create event button, added delete day function
